pyproject.toml
[tool.ruff]
# Exclude a variety of commonly ignored directories.
extend-exclude = [
    ".direnv",
    ".git",
    ".ruff_cache",
    "node_modules",
    "migrations",
]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
fixable = [
    "ALL",
]
unfixable = []
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.uv.sources]
# allow installing django-celery-beat 2.8.1 with django 6.0
# until https://github.com/celery/django-celery-beat/issues/977 is resolved
django-celery-beat = { git = "https://github.com/celery/django-celery-beat", rev = "bd429063cbb227c00905e7693c38895e44ff8e47" }

[tool.setuptools]
py-modules = []

[project]
name = "ch-manage2"
version = "0.1.0"
description = "Quick to Setup Church Management Software"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Django>=6.0",
    "boto3",
    "celery",
    "celery-progress",
    "celery[redis]",
    "channels",
    "channels_redis",
    "django-allauth[socialaccount,mfa]>=65.13.0",
    "django-anymail[mailgun]", # email services
    "django-celery-beat",
    "django-environ",
    "django-health-check",
    "django-hijack",
    "django-htmx",
    "django-recurrence>=1.14",
    "django-storages",
    "django-vite>=3.1.0",
    "django-waffle",
    "djangorestframework",
    "djangorestframework-api-key",
    "drf-spectacular",
    "httpx>=0.28.1",
    "markdown",
    "nh3>=0.2.21",
    "openai",
    "psycopg[binary]",
    "pydantic-ai>=1.0",
    "redis",
    "requests",
    "sentry-sdk[django]>=2.33.2",
    "whitenoise[brotli]",
]

[dependency-groups]
dev = [
    "channels[daphne]",
    "django-browser-reload",
    "pre-commit",
    "ruff>=0.13.1",
    "pegasus-cli>=0.12",
]
prod = [
    "gunicorn",
    "uvicorn[standard]",
]
